Moved new code to a new module
